RegisterSchema, GetLastMessageId

+++ SharpPulsar/Impl/conf/ClientConfigurationData.cs
@@ -30,7 +30,7 @@ namespace SharpPulsar.Impl.Conf
 	/// </summary>
 	public sealed class ClientConfigurationData
 	{
-		public long OperationTimeoutMs { get; set; } = 30000;
+		public int OperationTimeoutMs { get; set; } = 30000;
         public int WebServicePort { get; set; } = 8080;
         public string WebServiceScheme { get; set; } = "http";
 		public bool UseProxy { get; set; } = false;
